pyproject.toml
[project]
name = "maas-code-reviewer"
version = "0.1.0"
description = "An LLM-powered code reviewer for Launchpad merge proposals and GitHub pull requests. Built by the MAAS team at Canonical."
readme = "README.md"
authors = [
    { name = "MAAS Developers" }
]
requires-python = ">=3.12"
dependencies = [
    "google-genai",
    "keyring",
    "launchpadlib",
    "PyGithub",
]

[project.scripts]
maas-code-reviewer = "maas_code_reviewer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "ruff>=0.15.2",
    "ty>=0.0.1a7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.coverage.run]
source = ["src/maas_code_reviewer"]

[tool.coverage.report]
show_missing = true
skip_covered = true
